core-api: move LocalStorageFeatureFlags to api implementations
